[cpu] Support sending module logs to ElasticSearch and Grafana Loki
A new `<logs>` child element has been added to the `<cpu>` element
that can be configured similar to the following.

```
<ot-sim>
  <cpu>
    <logs>
      <elastic index="ot-sim-logs">http://1.2.3.4:9200</elastic>
      <loki>http://1.2.3.4:3100</loki>
    </logs>
  </cpu>
</ot-sim>
```

One of `elastic` or `loki` can be provided, or both can be provided.

Messages written to either location will contain document fields or
labels hinting at the hostname they came from, the module they're for,
and if they are a log or error message.