Node-RED support for OT-sim
This commit includes custom Node-RED nodes for interacting with the
OT-sim message bus. Node-RED on an OT-sim device can be enabled via the
Node-RED module, allowing users to connect to the Node-RED UI to
configure and deploy an HMI or use Node-RED as a replacement for the
existing OT-sim logic module.