feat(mqtt): support custom template for MQTT topic payloads
A `<payload-template>` element can be provided as part of the MQTT
client config. The value of the element should be a valid Golang text
template string, and can reference the following template variables:

  .Epoch     - int64 value representing number of seconds since epoch
  .Timestamp - string representation of current UTC time
  .Topic     - topic name
  .Value     - current value for topic

The `Timestamp` variable defaults to the Golang `time.RFC3339` format.
If a different format is needed, it can be specified using the
`timestamp` attribute for the `<payload-template>` element.

If no `<payload-template>` element is provided, the topic payload
defaults to the current value for the topic.

The following configuration example will generate a topic payload
similar to `2023-06-05 04:04:18, 3.45467, load-671/active_power`.

```
<mqtt>
  <endpoint>tcp://localhost:1883</endpoint>
  <client-id>ot-sim-test</client-id>
  <period>5s</period>
  <payload-template timestamp="2006-01-02 15:04:05">{{ .Timestamp }}, {{ .Value }}, {{ .Topic }}"
  <tag>load-671.voltage</tag>
  <tag>load-671.current</tag>
  <tag>load-671.active_power</tag>
  <tag>load-671.reactive_power</tag>
</mqtt>
```