feat: new Tailscale module
The `tailscale` module can be used to add an ot-sim device (bare metal,
VM, Docker container, etc) to a Tailscale tailnet at the underlying OS
level. It does so by shelling out to start the `tailscaled` daemon and
then shelling out to `tailscale up` to add the device to the tailnet.

A Tailscale auth key must be provided, either in the XML config or via
an environment variable. A hostname to use for the device on the tailnet
can also be provided. If not provided, the underlying OS hostname is
used.

```
<tailscale>
  <auth-key>tskey-auth-...</auth-key>
  <hostname>foobar</hostname>
  <accept-dns>false</accept-dns>
</tailscale>
```

The `OTSIM_TAILSCALE_AUTHKEY` environment variable can be used instead
of the `<auth-key>` configuration element if needed.

Note that the underlying OS (bare metal, VM, Docker container) must
already have the `tailscaled` and `tailscale` executables installed. See
the main OT-sim Dockerfile for an example of how to install the
Tailscale executables.

Also note that, in order for Tailscale to work in a Docker container,
the `NET_ADMIN` capability must be added and the `/dev/net/tun` and
`/lib/modules` volumes must be mounted (`/lib/modules` can be mounted
read-only).