feat(modbus): support various register data types
The Modbus server and client register configurations now support an
additional `data-type` attribute to specify what type of numerical data
type the register should represent. Valid data type values are as
follows:

  * int16
  * uint16
  * int32
  * uint32
  * int64
  * uint64
  * float
  * double

If not provided, the data type defaults to `uint16`.

Since the `data-type` attribute is only useful for numerical data types,
it is ignored if provided with registers of type `coil` or `discrete`.

Multiple Modbus registers are used to represent numerical data types
larger than 16 bits, and are combined using big endian byte order. It is
currently up to the user to ensure register addressing is correctly
configured to avoid register overlap for larger numerical data types. A
future release of the Modbus module will include validation code to
ensure there is no overlap.

Below is a configuration example:

```
<register type="holding" data-type="uint32">
  <address>40000</address>
  <tag>example.bigger_number</tag>
  <scaling>2</scaling>
</register>
<register type="holding" data-type="int16">
  <address>40002</address>
  <tag>example.smaller_number</tag>
  <scaling>1</scaling>
</register>
```